refactor(topology): move topology-related files to separate folders
